feat : 형변환 추가
24.11.11

+++ Simple2DGameEngine/DRenderer.cpp
@@ -64,16 +64,16 @@ void DRenderer::DrawBySpriteType(DOSprite* sprite)
 void DRenderer::DrawRectangel(DOSprite* sprite)
 {
     Reset();
-    MoveToEx(hdc, sprite->GetLeftTopPosition().x, sprite->GetLeftTopPosition().y, NULL);
-    LineTo(hdc, sprite->GetRightTopPosition().x, sprite->GetRightTopPosition().y);
-    LineTo(hdc, sprite->GetRightBottomPosiiton().x, sprite->GetRightBottomPosiiton().y);
-    LineTo(hdc, sprite->GetLeftBottomPosition().x, sprite->GetLeftBottomPosition().y);
-    LineTo(hdc, sprite->GetLeftTopPosition().x, sprite->GetLeftTopPosition().y);
+    MoveToEx(hdc, (int)sprite->GetLeftTopPosition().x, (int)sprite->GetLeftTopPosition().y, NULL);
+    LineTo(hdc, (int)sprite->GetRightTopPosition().x, (int)sprite->GetRightTopPosition().y);
+    LineTo(hdc, (int)sprite->GetRightBottomPosiiton().x, (int)sprite->GetRightBottomPosiiton().y);
+    LineTo(hdc, (int)sprite->GetLeftBottomPosition().x, (int)sprite->GetLeftBottomPosition().y);
+    LineTo(hdc, (int)sprite->GetLeftTopPosition().x, (int)sprite->GetLeftTopPosition().y);
 }
 
 void DRenderer::DrawEllipse(DOSprite* sprite)
 {
     Reset();
-    Ellipse(hdc, sprite->GetLeftTopPosition().x, sprite->GetLeftTopPosition().y, 
-        sprite->GetRightBottomPosiiton().x, sprite->GetRightBottomPosiiton().y);
+    Ellipse(hdc, (int)sprite->GetLeftTopPosition().x, (int)sprite->GetLeftTopPosition().y,
+        (int)sprite->GetRightBottomPosiiton().x, (int)sprite->GetRightBottomPosiiton().y);
 }
